Move blocking APIs out of commonMain
