+++ library/process/api/process.api
@@ -1,3 +1,16 @@
+public abstract class io/matthewnelson/kmp/process/Blocking {
+	public final fun waitFor ()I
+	public final fun waitFor-LRDsOJo (J)Ljava/lang/Integer;
+}
+
+public abstract class io/matthewnelson/kmp/process/Blocking$Waiter {
+	protected final field process Lio/matthewnelson/kmp/process/Process;
+	public synthetic fun <init> (Lio/matthewnelson/kmp/process/Process;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public final fun awaitStop ()Lio/matthewnelson/kmp/process/Process;
+	protected abstract fun isStarted ()Z
+	protected abstract fun isStopped ()Z
+}
+
 public abstract interface annotation class io/matthewnelson/kmp/process/InternalProcessApi : java/lang/annotation/Annotation {
 }
 
@@ -42,7 +55,7 @@ public abstract interface class io/matthewnelson/kmp/process/OutputFeed {
 	public abstract fun onOutput (Ljava/lang/String;)V
 }
 
-public abstract class io/matthewnelson/kmp/process/OutputFeed$Handler {
+public abstract class io/matthewnelson/kmp/process/OutputFeed$Handler : io/matthewnelson/kmp/process/Blocking {
 	protected field isDestroyed Z
 	public synthetic fun <init> (Lio/matthewnelson/kmp/process/Stdio$Config;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
 	protected final fun dispatchStderr (Ljava/lang/String;)V
@@ -59,12 +72,9 @@ public abstract class io/matthewnelson/kmp/process/OutputFeed$Handler {
 	public final fun stdoutWaiter ()Lio/matthewnelson/kmp/process/OutputFeed$Waiter;
 }
 
-public abstract class io/matthewnelson/kmp/process/OutputFeed$Waiter {
+public abstract class io/matthewnelson/kmp/process/OutputFeed$Waiter : io/matthewnelson/kmp/process/Blocking$Waiter {
 	public synthetic fun <init> (Lio/matthewnelson/kmp/process/Process;ZLkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public final fun awaitStop ()Lio/matthewnelson/kmp/process/Process;
 	public final fun awaitStopAsync (Lkotlin/jvm/functions/Function2;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	protected abstract fun isStarted ()Z
-	protected abstract fun isStopped ()Z
 }
 
 public abstract class io/matthewnelson/kmp/process/Process : io/matthewnelson/kmp/process/OutputFeed$Handler {
@@ -81,8 +91,6 @@ public abstract class io/matthewnelson/kmp/process/Process : io/matthewnelson/km
 	public final fun isAlive ()Z
 	public abstract fun pid ()I
 	public final fun toString ()Ljava/lang/String;
-	public final fun waitFor ()I
-	public final fun waitFor-LRDsOJo (J)Ljava/lang/Integer;
 	public final fun waitForAsync (Lkotlin/jvm/functions/Function2;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 	public final fun waitForAsync-KLykuaI (JLkotlin/jvm/functions/Function2;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 }
