+++ library/process/src/blockingMain/kotlin/io/matthewnelson/kmp/process/internal/-BlockingPlatform.kt
@@ -23,7 +23,12 @@ import io.matthewnelson.kmp.file.InterruptedException
 import io.matthewnelson.kmp.process.Output
 import io.matthewnelson.kmp.process.Signal
 import io.matthewnelson.kmp.process.Stdio
+import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
+
+@Suppress("NOTHING_TO_INLINE")
+@Throws(InterruptedException::class)
+internal expect inline fun Duration.threadSleep()
 
 @Throws(IOException::class)
 internal fun PlatformBuilder.blockingOutput(
