+++ library/process/src/blockingMain/kotlin/io/matthewnelson/kmp/process/internal/InputStream.kt
